feat: Enhance event handling and data structure in ROOT classes
- Added new 'Pointing' class to encapsulate array azimuth and altitude data.
- Updated 'DL2Event' class to include 'set_tel_estimate_disp' method for improved telescope displacement handling.
- Enhanced 'ExtraParameters' class with additional fields for better image parameter representation.
- Modified 'RootSimulatedCamera' and 'RootWriter' classes to support new simulated camera data handling.
- Improved 'GeometryReconstructor' with new methods for converting to field of view coordinates.
- Updated Python bindings to reflect changes in data structures and methods.

+++ include/histogram.hpp
@@ -717,7 +717,7 @@ Profile1D<Precision> make_profile(std::unique_ptr<Axis<Precision>> axis) {
 
 
 // Convenience functions for common histogram types
-template<typename Precision = double>
+template<typename Precision = float>
 Histogram1D<Precision> make_regular_histogram(Precision min, Precision max, int bins) {
     return make_histogram<Precision>(make_regular_axis<Precision>(min, max, bins));
 }
